Move function
This makes the function available to other jobs.
